requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
Booktype==1.5
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.1.3
django-fernet-encrypted-fields==0.3.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
et_xmlfile==2.0.0
frozenlist==1.7.0
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
multidict==6.6.4
openpyxl==3.1.5
packaging==25.0
pillow==11.0.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zope.interface==7.2
